the reviews are now correctly insert in review collection and user the movement between documentEmbedding and documentLinking has also been implemented

+++ src/main/java/it/unipi/CarRev/model/Review.java
@@ -21,5 +21,5 @@ public class Review {
     private Integer likes;
     private Integer report;
     private Integer year;
-    private Integer mileage;
+    private Double mileage;
 }
